Minor update-changes bugfix when initializing CHANGES via git tags
When running "update-changes -I" on a repo that already has git version tags, we
didn't correctly number the commit of the initial CHANGES entry. It's an
off-by-one, since the introduction of CHANGES itself would be the right
number. This would cause the introductory message to appear once more in a
subsequent regular "update-changes".

To work around this, the tag-derived introduction now first commits an empty
CHANGES, then derives the right version string, updates CHANGES once more, and
augments that existing commit.

Also a small whitespace fix to remove an erroneous extra empty line in CHANGES
when creating it.